# Final Project/docs/api.yaml
swagger: "2.0"
info:
  description: "Aplikasi ini digunakan untuk menyimpan foro maupun membuat comment untuk foto orang lain."
  version: "1.0.0"
  title: "MyGram"
  contact:
    email: "abdul.rasyid.anshori@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost"
basePath: "/v1"
...
schemes:
- "https"
- "http"
paths:
...
  /photos/{photoId}:
    put:
      tags:
      - "photo"
      summary: "Update Photo"
      description: "Endpoint ini digunakan untuk update photo."
      operationId: "updatePhoto"
      produces:
      - "application/json"
      parameters:
        - in: "header"
          name: "Authorization Bearer"
          type: "string"
          description: "Masukan Bearer Token yang diterima pada saat login"
          required: true
        - in: "path"
          name: "photoId"
          description: "ID photo yang akan di update"
          required: true
          type: "integer"
          format: "int64"
        - in: "body"
          name: "body"
          schema:
            $ref: "#/definitions/PhotoCreateAndUpdate"
      responses:
        "200":
          description: "OK"
          schema: 
            $ref: "#/definitions/PhotoUpdateResponse"
      security:
      - Bearer: []
    delete:
      tags:
      - "photo"
      summary: "Delete Photo"
      description: "Endpoint ini digunakan untuk menghapus photo."
      operationId: "deletePhoto"
      produces:
      - "application/json"
      parameters:
        - in: "header"
          name: "Authorization Bearer"
          type: "string"
          description: "Masukan Bearer Token yang diterima pada saat login"
          required: true
        - in: "path"
          name: "photoId"
          description: "ID photo yang akan di delete"
          required: true
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "OK"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Your photo has been successfully deleted"
      security:
      - Bearer: []
...
  /comments/{photoId}:
    put:
      tags:
      - "comment"
      summary: "Update Photo"
      description: "Endpoint ini digunakan untuk update photo."
      operationId: "updatePhoto"
      produces:
      - "application/json"
      parameters:
        - in: "header"
          name: "Authorization Bearer"
          type: "string"
          description: "Masukan Bearer Token yang diterima pada saat login"
          required: true
        - in: "path"
          name: "photoId"
          description: "ID photo yang akan di update"
          required: true
          type: "integer"
          format: "int64"
        - in: "body"
          name: "body"
          schema:
            $ref: "#/definitions/PhotoCreateAndUpdate"
      responses:
        "200":
          description: "OK"
          schema: 
            $ref: "#/definitions/PhotoUpdateResponse"
      security:
      - Bearer: []
    delete:
      tags:
      - "comment"
      summary: "Delete Photo"
      description: "Endpoint ini digunakan untuk menghapus photo."
      operationId: "deletePhoto"
      produces:
      - "application/json"
      parameters:
        - in: "header"
          name: "Authorization Bearer"
          type: "string"
          description: "Masukan Bearer Token yang diterima pada saat login"
          required: true
        - in: "path"
          name: "photoId"
          description: "ID photo yang akan di delete"
          required: true
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "OK"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Your photo has been successfully deleted"
      security:
      - Bearer: []
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  PhotoCreateAndUpdate:
    type: "object"
    properties:
      title:
        type: "string"
      caption:
        type: "string"
      photo_url:
        type: "string"
...
  PhotoUpdateResponse:
    allOf:
    - type: object
      properties:
        id:
          type: "integer"
    - $ref: "#/definitions/PhotoCreateAndUpdate"
    - type: object
      properties:
        user_id:
          type: "integer"
        updated_at:
          type: "string"
          example: "Date"